fix(network): IPv6: don't wait for RA for static IPv6 assignments
This patch reverts commit
d652354 ("wait for IPv6 RA if using none/static IPv6 assignment").
It's not generally correct wait for a default route to be established
for an interface, or to wait for "proto ra" routes in general.
For example, if the system is a router itself, it will receive no
RAs. In isolated networks, no gateway may be advertized, either.
This is similar in spirit to b5c4b5d ("Revert "wait for IPv6 RA
if using none/static IPv6 assignment"")

Whatever d652354 ("wait for IPv6 RA if using none/static IPv6 assignment")
was supposed to achieve, it should be done differently.