test: make package libdir configurable
Make the package libdir configurable to allow running the test suite
out-of-tree. `PKGLIBDIR` can be pointed to the installed dracut libdir.
Use `PKGLIBDIR` for the location of `dracut-init.sh`, `dracut-install`,
`dracut-util`, `dracut-cpio`, and `modules.d`.

This change is used to run the test suite against the installed dracut
in an autopkgtest for the Debian/Ubuntu package.

Signed-off-by: Benjamin Drung <benjamin.drung@canonical.com>