fix(systemd-repart): correct undefined $libdir
Nothing in dracut appears to define the '$libdir' environment
variable when module scripts are being run. As a result when
the 01-systemd-repart module script runs, it fails to copy any
of the config files into the initrd. Replace it with '/usr/lib'
which reflects where systemd-repart expects its global config
files.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>