test: skip test 50,60 if ifcfg dracut module can not be installed
Test 50 and 60 have a dependency on the ifcfg dracut module.
If ifcfg dracut module is not available, than the preconditions
for these tests are not met.

Since ifcfg dracut module is not available in many Linux enviroments
(such as Arch or Debian), it make sense to explicitelly skip these
tests instead of failing them as they are simply not applicable
to many Linux distributions in their current form.

Instead of disabling the tests, they should be reimplement
so that they do not require ifcfg dracut module in the future.