perf(systemd-udevd): remove duplicate rules
During a typical systemd-enabled initrd generation both udev-rules
and systemd-udevd gets installed. This leads to duplicate file copying,
that can be easily avoided just by removing the duplicate rules.

The performance benefit of this PR is negligent. More importantly
this PR makes it easier to read, reason and maintain the code.