perf(dracut-install): preload kmod resources for quicker module lookup
libkmod reads the `modules.*.bin` files in the kernel dir every time we
call one of the `kmod_module_new_from_lookup()` function. We end up
calling it a lot especially in ARM64 systems where we handle fw_devlink
supplier relations, which slows down things by a significant margin.

Instead, have it read these files at the start of the dracut-install run
so the module information can be looked up quicker. We don't need to
unload these resources in the cleanup function because libkmod handles
that itself in `kmod_unref()`.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>